Move flood detector to individual mod
