+++ nym-vpn-core/crates/nym-common/src/lib.rs
@@ -3,6 +3,8 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 mod error;
+pub mod flood;
+
 pub use error::*;
 
 #[cfg(target_os = "linux")]
